chore: Rename all remaining goose references to leaf
- Renamed .goosehints -> .leafhints
- Renamed goose-self-test.yaml -> leaf-self-test.yaml
- Renamed scripts/goose-db-helper.sh -> scripts/leaf-db-helper.sh
- Renamed crates/leaf/src/goose_apps/ -> crates/leaf/src/leaf_apps/
- Renamed goose_doc_guide.md -> leaf_doc_guide.md
- Renamed test snapshot files from goose__* to leaf__*
- Deleted goose.tar.gz build artifact
- Updated all code references from goose_apps to leaf_apps:
  - lib.rs: pub mod goose_apps -> pub mod leaf_apps
  - agents/platform_extensions/apps.rs
  - leaf_apps/cache.rs
  - leaf-server routes and openapi
- Fixed clippy warnings in configure.rs (sort_by -> sort_by_key)

All verification checks pass:
- cargo check: PASSED
- cargo clippy: PASSED (no warnings)
- cargo test: 160 passed
- cargo build --release: PASSED
