+++ crates/leaf-server/src/routes/agent.rs
@@ -11,7 +11,7 @@ use axum::{
     Json, Router,
 };
 use goose::agents::{Container, ExtensionLoadResult};
-use goose::goose_apps::{fetch_mcp_apps, GooseApp, McpAppCache};
+use leaf::leaf_apps::{fetch_mcp_apps, GooseApp, McpAppCache};
 
 use base64::Engine;
 use goose::agents::reply_parts::is_tool_visible_to_app;
