+++ crates/leaf/src/agents/builtin_skills/skills/leaf_doc_guide.md
