+++ crates/leaf/src/lib.rs
@@ -8,7 +8,7 @@ pub mod conversation;
 pub mod download_manager;
 pub mod execution;
 pub mod gateway;
-pub mod goose_apps;
+pub mod leaf_apps;
 pub mod hints;
 pub mod logging;
 pub mod mcp_utils;
